requirements.txt
fastmcp>=0.1.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
fastmcp-credentials>=0.1.0